+++ modal-go/image.go
@@ -12,18 +12,77 @@ import (
 type Image struct {
 	ImageId string
 
+	imageRegistryConfig *pb.ImageRegistryConfig
+	tag                 string
+
 	//lint:ignore U1000 may be used in future
 	ctx context.Context
 }
 
-func fromRegistryInternal(app *App, tag string, imageRegistryConfig *pb.ImageRegistryConfig) (*Image, error) {
+// NewImageFromRegistry builds a Modal Image from a public or private image registry without any changes.
+func NewImageFromRegistry(tag string, options *ImageFromRegistryOptions) *Image {
+	if options == nil {
+		options = &ImageFromRegistryOptions{}
+	}
+	var imageRegistryConfig *pb.ImageRegistryConfig
+	if options.Secret != nil {
+		imageRegistryConfig = pb.ImageRegistryConfig_builder{
+			RegistryAuthType: pb.RegistryAuthType_REGISTRY_AUTH_TYPE_STATIC_CREDS,
+			SecretId:         options.Secret.SecretId,
+		}.Build()
+	}
+
+	return &Image{
+		ImageId:             "",
+		imageRegistryConfig: imageRegistryConfig,
+		tag:                 tag,
+	}
+}
+
+// NewImageFromAwsEcr creates an Image from an AWS ECR tag.
+func NewImageFromAwsEcr(tag string, secret *Secret) *Image {
+	imageRegistryConfig := pb.ImageRegistryConfig_builder{
+		RegistryAuthType: pb.RegistryAuthType_REGISTRY_AUTH_TYPE_AWS,
+		SecretId:         secret.SecretId,
+	}.Build()
+
+	return &Image{
+		ImageId:             "",
+		imageRegistryConfig: imageRegistryConfig,
+		tag:                 tag,
+	}
+}
+
+// NewImageFromGcpArtifactRegistry creates an Image from a GCP Artifact Registry tag.
+func NewImageFromGcpArtifactRegistry(tag string, secret *Secret) *Image {
+	imageRegistryConfig := pb.ImageRegistryConfig_builder{
+		RegistryAuthType: pb.RegistryAuthType_REGISTRY_AUTH_TYPE_GCP,
+		SecretId:         secret.SecretId,
+	}.Build()
+	return &Image{
+		ImageId:             "",
+		imageRegistryConfig: imageRegistryConfig,
+		tag:                 tag,
+	}
+}
+
+func (image *Image) build(app *App) (*Image, error) {
+	if image == nil {
+		return nil, InvalidError{"image must be non-nil"}
+	}
+
+	// Image is already hyrdated
+	if image.ImageId != "" {
+		return image, nil
+	}
+
 	resp, err := client.ImageGetOrCreate(
 		app.ctx,
 		pb.ImageGetOrCreateRequest_builder{
 			AppId: app.AppId,
 			Image: pb.Image_builder{
-				DockerfileCommands:  []string{`FROM ` + tag},
-				ImageRegistryConfig: imageRegistryConfig,
+				DockerfileCommands:  []string{`FROM ` + image.tag},
+				ImageRegistryConfig: image.imageRegistryConfig,
 			}.Build(),
 			BuilderVersion: imageBuilderVersion(""),
 		}.Build(),
@@ -86,9 +145,7 @@ func fromRegistryInternal(app *App, tag string, imageRegistryConfig *pb.ImageReg
 		return nil, RemoteError{fmt.Sprintf("Image build for %s failed with unknown status: %s", resp.GetImageId(), result.GetStatus())}
 	}
 
-	img := &Image{
-		ImageId: resp.GetImageId(),
-		ctx:     app.ctx,
-	}
-	return img, nil
+	image.ImageId = resp.GetImageId()
+	image.ctx = app.ctx
+	return image, nil
 }