+++ modal-go/test/sandbox_filesystem_snapshot_test.go
@@ -1,3 +1,4 @@
+//nolint:staticcheck // SA1019 We need to use deprecated API for testing
 package test
 
 import (